IMGAPI-391: use hostnames (we have DNS) so don't need to ignore host certs
Also use 'close' instead of 'end' to watch for end of reading
a file for sha1 calc. Very rarely (1% or less) we see a test failure
on sha1 mismatch (see below). A theory is that we get a 'data'
event after the 'end'. I don't know that is true tho.

```
not ok 202 sha1 matches image data
 ---
   type:    AssertionError
   message: sha1 matches image data
   code:    ~
   errno:   ~
   file:    /opt/smartdc/imgapi/test/image-basics.dc-test.js
   line:    315
   column:  19
   stack:
     - ReadStream.<anonymous> (/opt/smartdc/imgapi/test/image-basics.dc-test.js:315:19)
     - ReadStream.EventEmitter.emit (events.js:117:20)
     - _stream_readable.js:920:16
     - process._tickCallback (node.js:415:13)
   wanted:  da39a3ee5e6b4b0d3255bfef95601890afd80709
   found:   1b4240b257a89a190fd1a3c9eccffc7ccc951352
   diff:    |
     FOUND:  1b4240b257a89a190fd1a3c9eccffc7ccc951352
     WANTED: da39a3ee5e6b4b0d3255bfef95601890afd80709
             ^ (at position = 0)
```