Logic to deal with caps in image names
On newer images we switched to using proper caps for distro names in
the /etc/product file (e.g., "CentOS" instead of "centos" for example).
This change converts the image name to all lower case in oder to handle
any differences in capitalization of image name from one release to the
next.